requirements.txt
piper-tts
pymupdf
pygame
Pillow
PyQt6
